MAINT: Remove whitespace special path and fix newline loadtxt incompat
The `ignore_leading_spaces` is sufficient to handle everything that
the whitespace loader should be doing.  Although, the whitespace
version did allow empty lines (with or without whitespace), while
the other version did not.  Just align both, a new option could maybe
be implemented to not allow it if necessary.