MAINT,ENH: Use Python errors directly and improve messages
This changes the error on parsing to `ValueError` in most cases.
This is something we should potentially discuss.

If we use this function as a backend for multiple entry-points
some of the "pass error out" logic would maybe be useful to
give better context in the error message (e.g. if we had two
functions with different name for `usecols`).

For now, I think I prefer this, it feels simpler to me.