ENH,BUG: Fix generic "whitespace" delimiter
This makes the generic whitespace delimiter work the same as
the general `line.split()`.  Slightly bloats up the tokenizer,
but not too much.

The commit also includes a bit of cleanup for the tokenizer state
to be an enum (and I never actually used the weird "in field"
information, it was a bit of a misdesign in the state).