MAINT: Remove block manager
This commit removes the block manager.  This basically assumes
that the kernel is at least as good a block manager as we are.
I think this assumption is correct.  Timing does not suggest any
significant changes in either direction, the allocation/realloc
sizes could potentially be tuned.
I expect that if anything, this is very mildly faster sometimes
because we avoid the "copy to contiguous".