BUG: Simplify allocations and ensure we always allocate 1 byte
This could be simplified a bit more, and NumPy actually overallocates
by more normally.  But 1-byte allocation (just to have enough) should
be fine.

This also fixes the fact that we did not init the array correctly
when `max_rows` was used.

Adds tests for both an empty dtype (structured) and object dtype
with `max_rows`.

(And includes some minor maintanence I did not bother to split out)