BUG,MAINT: Fixup parsing of control characters
I currently made the choice of using -1 (this is unsigned so it is
actually the maximum value) as "not applicable" character.
Unicode stops at something like 21-bits, so it should be impossible
for us to hit this.  We could also use the NUL character and live
with the fact that embedded NUL's are weird and thus do weird things
(e.g. the Python parser just barfs on it).