MAINT,BUG: Work arrays into the core of rows
While working with arrays in the row reader is a bit annoying in
parts (because why?).  It also is somewhat necessary unfortunately.

This solves the problem that having the array already ready
greatly simplifies cleanup when the result has an object dtype.

(It also is a bit nicer for growing the strings, since we can just
use the NumPy API to copy.)

I don't think this is really much nicer and if NumPy grows better
API to decref a chunk of memory content, rather than relying on
having the array object around it would be worse probably...

But this is what we got currently, mutating/creating the array
object inside `read_rows` just seems more practical

This adds a test to check that cleanup is indeed successfull
(at least for the more simple cases).