MAINT: Remove `unistd.h` include and raw file streamer
This is necessary for windows support. The raw file streamer is
removed because it is outdated and not used for a longer time.

We could add a new raw-file reader at some point, but it seems not
a high priority, we should start with adding a chunked Python file
reader probably.