ENH: Chunk opened file-likes and otherwise use iteration
This reads file-likes we open in chunks and modifies iterables
to be read by iterating through them.
(If we use the comment pre-processing we always iterate)

Both improves performance, and cleans up things slightly, although
at the cost of having two "streamers" on the C-side that do fairly
similar things.