MAINT: Remove analyzer completely
This commit removes the analyzer and (most) related functionality.
This functionality is needed to replicate some of the more nifty
features of `genfromtxt` or `pandas`.

However, it is not necessary for `loadtxt` and currently relied
fairly heavly on a two-pass approach, which is not trivially compatible
with `genfromtext` (or `loadtxt`) to begin with.

The code may be interesting, so hopefully will be discoverable, e.g.
through the git history.

For someone looking into this, there may however be an alternative
approach more similar to `genfromtxt`:

Try parsing each field, but fail if the dtype is not compatible
(or check whether it is).  If it is not, stop parsing that row,
start a new array, and upgrade the dtype.  Then continue
with a new array.  At the end concatenate all arrays.

In principle it may even work to move that "try/except" logic into
Python, so long the C-code raises a specific exception.
(A problem is that the partially written line must be retried
though).