MAINT: Clean up encoding handling slightly
In particular, just use the `char *` in C directly, that seems easier.

This does not figure out whether the encoding we currently use is
actually the *right* encoding.
(I am also not 100% sure if encoding names can include unicode characters
and if utf-8 encoding of the encoding is right then :).  But I would be
extremely surprised if this is not right and even more if it mattered.)