MAINT: Remove leftover of discovery and old float-parsing tries
This second float parsing code is probably interesting for a
non-roundtripping version (my guess is equivalent to pandas high
precision?).  But I doubt we want to go there, and even if, it should
be easy enough to find it again (or steal it from pandas, etc.)

The other functions were still around because I never replaced them
in the analyzer.  Now that the analyzer is gone, we can delete them.