BUG: Catch all unicode "whitespaces" when parsing numbers
We could discuss rolling some of this back, but Python allows any
whitespace in numbers (e.g. even newlines if this is a quoted field).
This will probably slow parsing down an inkling, so if we want we
may be able to recover some of that.

But for now, it seems like the correct/safe thing to do.