BUG: Swap `\n` and `\r` in the `\r\n` support
I just had it wrong in most places, but since all that typically
happens would be inserting a blank line, it is hard to notice.

I had a fix already, but that actually made it possible to modify
the newline (and carriage) character (useful for `linterminator`
support other than universal-newline, also potentially useful to
ignore newline characters when the user passes a list of strings).

In any case, since we don't need that right now, this just swaps
fixes the logic for now.