MAINT: Simplify field-types code
The field-types are somewhat a relict of the dtype discovery
(and the string length discovery which was independent).
Both is gone now, so it is easier to simplify and consolidate the
code.

In the long run (i.e. as part of putting it into NumPy), we should
likely move the parser function onto the DType so that the field
types would be nothing more but a storage for faster access.